Initial refactor task base
